requirements.txt
yfinance==0.2.36
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
boto3==1.34.69
reportlab==4.1.0
Pillow==10.3.0
feedparser==6.0.11
python-dateutil==2.9.0
pytz==2024.1
finnhub-python==2.4.19
